Split ProfileDataService into individual services. Split tests.
