+++ aspnetcore/src/api/Services/Profiledata/IPublicationService.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using api.Models.ProfileEditor.Items;
+
+namespace api.Services.Profiledata
+{
+    public interface IPublicationService
+    {
+        Task<List<ProfileEditorPublication>> GetProfileEditorPublications(int userprofileId, bool forElasticsearch = false);
+    }
+}
