+++ aspnetcore/src/api/Services/Profiledata/ITelephoneNumberService.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using api.Models.ProfileEditor.Items;
+
+namespace api.Services.Profiledata
+{
+    public interface ITelephoneNumberService
+    {
+        Task<List<ProfileEditorTelephoneNumber>> GetProfileEditorTelephoneNumbers(int userprofileId, bool forElasticsearch = false);
+
+    }
+}
